v5.9.8

+++ bus-office/src/main/java/org/aoju/bus/office/support/excel/ExcelKit.java
@@ -44,7 +44,7 @@ import java.io.OutputStream;
  * Excel工具类
  *
  * @author Kimi Liu
- * @version 5.9.6
+ * @version 5.9.8
  * @since JDK 1.8+
  */
 public class ExcelKit {
@@ -548,7 +548,7 @@ public class ExcelKit {
             }
             int remainder = index % 26;
             colName.append((char) (remainder + 'A'));
-            index = (int) ((index - remainder) / 26);
+            index = (index - remainder) / 26;
         } while (index > 0);
         return colName.reverse().toString();
     }
